+++ Money_Support/components/fullComp/expendituresIncomistItem.jsx
@@ -1,9 +1,10 @@
 import React from "react";
-import { Dimensions, StyleSheet, View } from "react-native";
+import { Dimensions, Pressable, StyleSheet, View } from "react-native";
 import colorPallet from "../../constants/Colors";
 import DefaultLabel from "../Lables/default";
 import ValueToString from "../../Functions/valueToString";
 import FontAwsomeIcon from "react-native-vector-icons/FontAwesome"
+import EditEntryModal from "../Modals/editEntryModal";
 
 const ExpendituresIncomListItem = (props) => {
     const screenWidth = Dimensions.get("window").width
@@ -31,13 +32,14 @@ const ExpendituresIncomListItem = (props) => {
     }
 
     return(
-        <View style={styles.parantView} >    
+        <Pressable style={styles.parantView} onLongPress={() => props.openEditentry(props.dataID)} delayLongPress={1000}>    
             <View style={{backgroundColor:bgColor, width:30, height:25, marginTop:5, marginLeft:5, borderRadius:5}}>
                 <FontAwsomeIcon name={`${props.status == "Expenditures" ? "arrow-down": props.status == "Income" ? "arrow-up": "arrow-right"}`} size={20} marginLeft={6} marginTop={2} color={colorPallet.bg_2e}/>
             </View>
             <DefaultLabel text={props.date} width={120} marginLeft={5} marginTop={5} marginBottom={5} backGround={colorPallet.bg_4e}/>
             <DefaultLabel text={`${props.status == "Expenditures" ? "-": props.status == "Income" ? "+": ""}${ValueToString(props.value)}`} width={screenWidth-220} marginLeft={5} marginTop={5} marginBottom={5} backGround={colorPallet.bg_4e}/>
-        </View>
+        </Pressable>
+
     );
 }