+++ Money_Support/variables/string.js
@@ -44,5 +44,5 @@ export const setCurrentUserKey = (value) => {
     currentuserKey = value
 }
 
-export const appVersion = "0.2.3 - Early Acess"
+export const appVersion = "0.3.0 - Early Acess"
 export const expoVersion = "52.0.32"